small folder structure refactoring, concerning node registry
