+++ mipengine/node_registry.py
