+++ tests/integration_tests/test_node_registry.py
@@ -1,5 +1,5 @@
 import pytest
-from mipengine.node_registry.node_registry import (
+from mipengine.node_registry import (
     NodeRegistryClient,
     NodeRole,
     Pathology,
